Implement timerlock feature for timerstallers
This is a punishment that can be used against timerstallers.  Basically, if a player is caught to be timerstalling, a staff can verbally warn them to stop, and if they continue, timerlock them.

Timerlocks will make it so that that player only has 3 timer ticks left (30 seconds) for the current battle, and then will block them from laddering for 7 days.