Add [Gen 1] Random Auto Level Adjusted format
This change adds a new format which is very similar to [Gen1] Random Battle. The main difference is that Pokemon levels are adjusted automatically so that, over time, each pokemon contributes the same amount to a team on average.

The file config/pokemon-match-records.tsv is used to store records for matches played under this format. For each pokemon, it stores how many times teams with that pokemon won or lost, and its current adjusted level. All pokemon are set to level 50 initially.

Whenever a match ends, we update the record and evaluate whether any pokemon should be adjusted to be higher or lower level: pokemon whose teams win over half the time are scaled down, and vice versa. We do some probabilistic magic to make sure that this happens gradually and sensibly and that we do not overadjust.

The code represents the contents of this TSV file as an object called pokemonRecord which is updated and passed around various pieces of code as necessary.