Add [Gen 1] Random Auto Level Adjusted format
This change adds a new format which is very similar to [Gen1] Random Battle. The main difference is that Pokemon levels are adjusted automatically so that, over time, each pokemon contributes the same amount to a team on average.

The file config/pokemon-match-records.tsv is used to store records for matches played under this format. For each pokemon, it stores how many times teams with that pokemon won or lost, and its current adjusted level. All pokemon are set to level 50 initially.

Whenever a match ends, we update the record and evaluate whether any pokemon should be adjusted to be higher or lower level: pokemon whose teams win over half the time are scaled down, and vice versa. We do some probabilistic magic to make sure that this happens gradually and sensibly and that we do not overadjust.

The code represents the contents of this TSV file as an object called pokemonRecord which is updated and passed around various pieces of code as necessary.

Gen 1 Random Auto Adjust: Create match records

For the new [Gen 1] Random Auto Level Adjusted format, this commit adds the initial version of the pokemon match records, so that each pokemon is recorded as not having played any matches and is set at level 50.

Gen 1 Random Auto Adjust: Tweak math, remove log()

This commit contains two minor tweaks:
1) Adjusting the random factor that determines whether to auto-adjust a pokemon's level based on performance. The old chance to adjust was 1/numMatches and the new one is 1/sqrt(numMatches), which will help us not get stuck in situations where a pokemon's adjusted level is off for an inordinate amount of time. So if a pokemon has been in 100 matches and its win/loss ratio is uneven, we now have a 1/10 chance to adjust its level instead of 1/100.
2) Removing some console.log statements I'd accidentally left in from debugging.