Create new [Gen 1] Random Auto Level Adjusted format (#220)
This format is very similar to the existing [Gen 1] Random format, except that all pokemon are initially at
level 50 and their levels are adjusted based on whether they win or lose in matches over time. So e.g. if
teams with Caterpie on them lose more often, Caterpie's level will be adjusted to be higher. As many matches are
played, each pokemon's level will eventually settle at a level so that it is as advantageous to have that
pokemon on a team as any other pokemon.
These match records are stored in the file config/pokemon-match-records.tsv - this records how many times each
pokemon has contributed to a win or loss, and its current adjusted level. This file does not exist initially -
it is created after the first battle in the new format, and updated after each subsequent battle. Its initial
state is stored in config/pokemon-match-records-initial.tsv. The code keeps track of the pokemon records by
passing around an object called pokemonRecord which is created by parsing the TSV file in room-battle.js.

Fix npm issue

Move new auto-adjusted format to Custom Gold Tiers